Cleanup handler chain
