+++ pkg/apiserver/handlers_test.go
@@ -17,19 +17,12 @@ limitations under the License.
 package apiserver
 
 import (
-	"fmt"
-	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"reflect"
-	"regexp"
-	"strings"
-	"sync"
 	"testing"
-	"time"
 
 	"k8s.io/kubernetes/pkg/api"
-	"k8s.io/kubernetes/pkg/api/errors"
 	"k8s.io/kubernetes/pkg/api/testapi"
 	"k8s.io/kubernetes/pkg/apis/extensions"
 	"k8s.io/kubernetes/pkg/auth/authorizer"
@@ -42,128 +35,12 @@ func (fakeRL) Stop()             {}
 func (f fakeRL) TryAccept() bool { return bool(f) }
 func (f fakeRL) Accept()         {}
 
-func expectHTTP(url string, code int) error {
-	r, err := http.Get(url)
-	if err != nil {
-		return fmt.Errorf("unexpected error: %v", err)
-	}
-	if r.StatusCode != code {
-		return fmt.Errorf("unexpected response: %v", r.StatusCode)
-	}
-	return nil
-}
-
 func getPath(resource, namespace, name string) string {
 	return testapi.Default.ResourcePath(resource, namespace, name)
 }
 
 func pathWithPrefix(prefix, resource, namespace, name string) string {
 	return testapi.Default.ResourcePathWithPrefix(prefix, resource, namespace, name)
-}
-
-// Tests that MaxInFlightLimit works, i.e.
-// - "long" requests such as proxy or watch, identified by regexp are not accounted despite
-//   hanging for the long time,
-// - "short" requests are correctly accounted, i.e. there can be only size of channel passed to the
-//   constructor in flight at any given moment,
-// - subsequent "short" requests are rejected instantly with appropriate error,
-// - subsequent "long" requests are handled normally,
-// - we correctly recover after some "short" requests finish, i.e. we can process new ones.
-func TestMaxInFlight(t *testing.T) {
-	const AllowedInflightRequestsNo = 3
-	// Size of inflightRequestsChannel determines how many concurrent inflight requests
-	// are allowed.
-	inflightRequestsChannel := make(chan bool, AllowedInflightRequestsNo)
-	// notAccountedPathsRegexp specifies paths requests to which we don't account into
-	// requests in flight.
-	notAccountedPathsRegexp := regexp.MustCompile(".*\\/watch")
-	longRunningRequestCheck := BasicLongRunningRequestCheck(notAccountedPathsRegexp, map[string]string{"watch": "true"})
-
-	// Calls is used to wait until all server calls are received. We are sending
-	// AllowedInflightRequestsNo of 'long' not-accounted requests and the same number of
-	// 'short' accounted ones.
-	calls := &sync.WaitGroup{}
-	calls.Add(AllowedInflightRequestsNo * 2)
-
-	// Responses is used to wait until all responses are
-	// received. This prevents some async requests getting EOF
-	// errors from prematurely closing the server
-	responses := sync.WaitGroup{}
-	responses.Add(AllowedInflightRequestsNo * 2)
-
-	// Block is used to keep requests in flight for as long as we need to. All requests will
-	// be unblocked at the same time.
-	block := sync.WaitGroup{}
-	block.Add(1)
-
-	server := httptest.NewServer(
-		MaxInFlightLimit(
-			inflightRequestsChannel,
-			longRunningRequestCheck,
-			http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				// A short, accounted request that does not wait for block WaitGroup.
-				if strings.Contains(r.URL.Path, "dontwait") {
-					return
-				}
-				if calls != nil {
-					calls.Done()
-				}
-				block.Wait()
-			}),
-		),
-	)
-	defer server.Close()
-
-	// These should hang, but not affect accounting.  use a query param match
-	for i := 0; i < AllowedInflightRequestsNo; i++ {
-		// These should hang waiting on block...
-		go func() {
-			if err := expectHTTP(server.URL+"/foo/bar?watch=true", http.StatusOK); err != nil {
-				t.Error(err)
-			}
-			responses.Done()
-		}()
-	}
-	// Check that sever is not saturated by not-accounted calls
-	if err := expectHTTP(server.URL+"/dontwait", http.StatusOK); err != nil {
-		t.Error(err)
-	}
-
-	// These should hang and be accounted, i.e. saturate the server
-	for i := 0; i < AllowedInflightRequestsNo; i++ {
-		// These should hang waiting on block...
-		go func() {
-			if err := expectHTTP(server.URL, http.StatusOK); err != nil {
-				t.Error(err)
-			}
-			responses.Done()
-		}()
-	}
-	// We wait for all calls to be received by the server
-	calls.Wait()
-	// Disable calls notifications in the server
-	calls = nil
-
-	// Do this multiple times to show that it rate limit rejected requests don't block.
-	for i := 0; i < 2; i++ {
-		if err := expectHTTP(server.URL, errors.StatusTooManyRequests); err != nil {
-			t.Error(err)
-		}
-	}
-	// Validate that non-accounted URLs still work.  use a path regex match
-	if err := expectHTTP(server.URL+"/dontwait/watch", http.StatusOK); err != nil {
-		t.Error(err)
-	}
-
-	// Let all hanging requests finish
-	block.Done()
-
-	// Show that we recover from being blocked up.
-	// Too avoid flakyness we need to wait until at least one of the requests really finishes.
-	responses.Wait()
-	if err := expectHTTP(server.URL, http.StatusOK); err != nil {
-		t.Error(err)
-	}
 }
 
 func TestReadOnly(t *testing.T) {
@@ -181,62 +58,6 @@ func TestReadOnly(t *testing.T) {
 			t.Fatalf("Couldn't make request: %v", err)
 		}
 		http.DefaultClient.Do(req)
-	}
-}
-
-func TestTimeout(t *testing.T) {
-	sendResponse := make(chan struct{}, 1)
-	writeErrors := make(chan error, 1)
-	timeout := make(chan time.Time, 1)
-	resp := "test response"
-	timeoutResp := "test timeout"
-
-	ts := httptest.NewServer(TimeoutHandler(http.HandlerFunc(
-		func(w http.ResponseWriter, r *http.Request) {
-			<-sendResponse
-			_, err := w.Write([]byte(resp))
-			writeErrors <- err
-		}),
-		func(*http.Request) (<-chan time.Time, string) {
-			return timeout, timeoutResp
-		}))
-	defer ts.Close()
-
-	// No timeouts
-	sendResponse <- struct{}{}
-	res, err := http.Get(ts.URL)
-	if err != nil {
-		t.Error(err)
-	}
-	if res.StatusCode != http.StatusOK {
-		t.Errorf("got res.StatusCode %d; expected %d", res.StatusCode, http.StatusOK)
-	}
-	body, _ := ioutil.ReadAll(res.Body)
-	if string(body) != resp {
-		t.Errorf("got body %q; expected %q", string(body), resp)
-	}
-	if err := <-writeErrors; err != nil {
-		t.Errorf("got unexpected Write error on first request: %v", err)
-	}
-
-	// Times out
-	timeout <- time.Time{}
-	res, err = http.Get(ts.URL)
-	if err != nil {
-		t.Error(err)
-	}
-	if res.StatusCode != http.StatusGatewayTimeout {
-		t.Errorf("got res.StatusCode %d; expected %d", res.StatusCode, http.StatusServiceUnavailable)
-	}
-	body, _ = ioutil.ReadAll(res.Body)
-	if string(body) != timeoutResp {
-		t.Errorf("got body %q; expected %q", string(body), timeoutResp)
-	}
-
-	// Now try to send a response
-	sendResponse <- struct{}{}
-	if err := <-writeErrors; err != http.ErrHandlerTimeout {
-		t.Errorf("got Write error of %v; expected %v", err, http.ErrHandlerTimeout)
 	}
 }
