+++ pkg/genericapiserver/routes/swaggerui.go
@@ -29,6 +29,7 @@ import (
 // SwaggerUI exposes files in third_party/swagger-ui/ under /swagger-ui.
 type SwaggerUI struct{}
 
+// Install adds the SwaggerUI webservice to the given mux.
 func (l SwaggerUI) Install(mux *apiserver.PathRecorderMux, c *restful.Container) {
 	fileServer := http.FileServer(&assetfs.AssetFS{
 		Asset:    swagger.Asset,
