+++ pkg/genericapiserver/routes/version.go
@@ -25,9 +25,10 @@ import (
 	"k8s.io/kubernetes/pkg/version"
 )
 
+// Version provides a webservice with version information.
 type Version struct{}
 
-// InstallVersionHandler registers the APIServer's `/version` handler
+// Install registers the APIServer's `/version` handler.
 func (v Version) Install(mux *apiserver.PathRecorderMux, c *restful.Container) {
 	// Set up a service to return the git code version.
 	versionWS := new(restful.WebService)
